cleand app.js, move some func to misc.js
